requirements.txt
asgiref==3.8.1
cffi==1.17.0
cryptography==43.0.0
Django==5.0.7
pillow==10.4.0
pycparser==2.22
PyMySQL==1.1.1
sqlparse==0.5.1
tzdata==2024.1
gunicorn
whitenoise

